+++ src/dataflow/multiwoz/create_belief_state_prediction_report.py
@@ -11,7 +11,7 @@ import json
 from dataclasses import dataclass
 from typing import Any, Dict, List, Union
 
-from dataflow.core.io import load_jsonl_file_and_build_lookup, save_jsonl_file
+from dataflow.core.io_utils import load_jsonl_file_and_build_lookup, save_jsonl_file
 from dataflow.core.prediction_report import (
     PredictionReportDatum,
     save_prediction_report_tsv,
